+++ src/java/businesslogic/Pants.java
@@ -19,7 +19,7 @@ public class Pants extends Item{
         return PantsDB.searchItems(group, con);
     }
         
-    public Pants(int id, String name, int price, int stock) {
+    public Pants(int id, String name, double price, int stock) {
         super(id, name, price, stock);
     }
